Fix broken xUnit 3 explicit sender support
